% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rStockData.R
\name{getYahooStockData}
\alias{getYahooStockData}
\title{Gets historical stock data from Yahoo Finance}
\usage{
getYahooStockData(symbol, endMon, endDay, endYear, startMon = 1,
  startDay = 1, startYear = 2010, cacheDir = "stockDataCache",
  useCache = FALSE)
}
\arguments{
\item{symbol}{Yahoo stock ticker symbol}

\item{endMon}{ending month}

\item{endDay}{ending day}

\item{endYear}{ending year}

\item{startMon}{start month}

\item{startDay}{start day}

\item{startYear}{start year}

\item{cacheDir}{cache directory}

\item{useCache}{use cached data if available}
}
\value{
a data frame of historical stock data
}
\description{
Gets historical stock data from Yahoo Finance
}
\details{
Based on a stock symbol and start and end dates, historical stock data is retrieved
 from Yahoo Finance.  If the start date isn't supplied, the default start date of
 1/1/2010 will be used.  If the useCache is true, a cached copy of the data will
 be saved the first time the data is retrieved.  Following calls will use the data
 from the cacheDir.

 The returned data frame will have the following columns:
 Date   Open   High    Low  Close   Volume Adj.Close
}
\examples{
# get data for Apple from 1/1/2010 - 1/1/2015
 getYahooStockData('AAPL', 1, 1, 2015)
}
\author{
Phillip Chin
}

